ml/requirements.txt
pandas
numpy
scikit-learn
SQLAlchemy
pymysql
python-dotenv
